python-sdk/requirement.txt
python-consul==1.1.0
pytest==7.1.3
PyYaml==6.0